Iterators 2.0
Collections with size greater than but not equal to the system vector width are
now handled entirely with SIMD operations, and work with swizzle-like
operations. It's up to the user to ensure their swizzles work with their
specific collection size (e.g. .flip() requires an even number of elements to
work as expected)